:art: :: PlayerView, MiniPlayerView로 분할
