+++ Projects/Features/PlayerFeature/Sources/MiniPlayerView.swift
@@ -1,0 +1,26 @@
+//
+//  MiniPlayerView.swift
+//  PlayerFeature
+//
+//  Created by YoungK on 2023/01/12.
+//  Copyright © 2023 yongbeomkwak. All rights reserved.
+//
+
+import UIKit
+import SwiftUI
+import Utility
+import DesignSystem
+import SnapKit
+import Then
+
+final class MiniPlayerView: UIView {
+    override init(frame: CGRect) {
+            super.init(frame: frame)
+        }
+        
+        required init?(coder: NSCoder) {
+            super.init(coder: coder)
+        }
+        
+        
+}
